Split logger abstraction
