+++ pkg/abstraction/logger.go
@@ -1,0 +1,16 @@
+package abstraction
+
+type LoggerName string
+
+type LoggerRecord interface {
+	Name() LoggerName
+}
+
+type LoggerRequest interface {
+	Name() LoggerName
+}
+
+type Logger interface {
+	PushRecord(LoggerRecord) error
+	PullRecord(LoggerRequest) (LoggerRecord, error)
+}
